refactor update package structure for profile-related components
