+++ feature/profile/profileUi/src/main/java/com/feature/profile/profileUi/screen/components/AppRestrictionDialog.kt
@@ -17,7 +17,7 @@ import androidx.compose.ui.graphics.Brush
 import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.unit.dp
 import com.feature.profile.profileUi.R
-import com.feature.profile.profileUi.screen.ContentRestriction
+import com.feature.profile.profileUi.screen.profile.ContentRestriction
 import com.paris.aflami.designsystem.components.AppDialog
 import com.paris.aflami.designsystem.components.ButtonType
 import com.paris.aflami.designsystem.components.CustomButton
