+++ feature/profile/profileUi/src/main/java/com/feature/profile/profileUi/screen/components/AppThemeDialog.kt
@@ -18,7 +18,7 @@ import androidx.compose.ui.graphics.Brush
 import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.unit.dp
 import com.feature.profile.profileUi.R
-import com.feature.profile.profileUi.screen.Appearance
+import com.feature.profile.profileUi.screen.profile.Appearance
 import com.paris.aflami.designsystem.components.AppDialog
 import com.paris.aflami.designsystem.components.ButtonType
 import com.paris.aflami.designsystem.components.CustomButton
